requirements.txt
# FastAPI Backend Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
google-cloud-aiplatform>=1.38.0
google-cloud-dialogflow-cx>=1.30.0
google-cloud-storage>=2.14.0
google-auth>=2.23.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
httpx>=0.26.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

